Refactor repository interfaces to use Integer as the ID type and add assignment methods for EmployeeProject

+++ src/main/java/org/ahmet/mysqlkafka/repository/ProjectRepository.java
@@ -3,5 +3,5 @@ package org.ahmet.mysqlkafka.repository;
 import org.ahmet.mysqlkafka.model.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjectRepository extends JpaRepository<Project, Long> {
+public interface ProjectRepository extends JpaRepository<Project, Integer> {
 }
